# [opt](column) Optimize the execution of the append_data_by_selector function in ColumnNullable.